requirement.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymysql==1.1.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0